+++ packages/node/src/indexer/dictionary/v1/substrateDictionaryV1.spec.ts
@@ -84,6 +84,12 @@ const eventHandler: SubstrateEventHandler = {
   filter: { method: 'event', module: 'module' },
 };
 
+const callHandlerWithUndefined: SubstrateCallHandler = {
+  kind: SubstrateHandlerKind.Call,
+  handler: 'handleCall',
+  filter: { isSigned: true, module: undefined, method: undefined },
+};
+
 describe('Building dictionary query entries', () => {
   it('supports block handlers', () => {
     /* If there are any blockhandlers without a modulo or timestamp filter we expect no query entries */
@@ -210,6 +216,24 @@ describe('Building dictionary query entries', () => {
           { field: 'filter1', value: 'foo' },
           { field: 'filter2', value: 'bar' },
         ],
+      },
+    ]);
+  });
+
+  it('create dictionary call filter condition and remove undefined fields', () => {
+    const result1 = buildDictionaryV1QueryEntries(
+      [makeDs([callHandlerWithUndefined])],
+      () => undefined,
+    );
+    expect(result1).toEqual([
+      {
+        conditions: [
+          {
+            field: 'isSigned',
+            value: true,
+          },
+        ],
+        entity: 'extrinsics',
       },
     ]);
   });