+++ packages/node/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Update with common-substrate, fix `EventFilter` incorrectly extend `BlockFilter`, lead dictionary error
+- Fix dictionary query entries included undefined fields
+
 ## [4.6.5] - 2024-06-18
 ### Fixed
 - Add missing chainType validation for local mode